Fix eslint errrs
