+++ src/components/InputTodo.js
@@ -1,10 +1,10 @@
-import React, { useState } from "react";
-import { FaPlusCircle  } from "react-icons/fa"
-import './styles/app.css'
-
+import React, { useState } from 'react';
+import { FaPlusCircle } from 'react-icons/fa';
+import './styles/app.css';
+/* eslint-disable react/prop-types */
 const InputTodo = ({ addTodoItem }) => {
-  const [title, setTitle] = useState("");
-  const [message, setMessage] = useState("");
+  const [title, setTitle] = useState('');
+  const [message, setMessage] = useState('');
 
   const handleChange = (e) => {
     setTitle(e.target.value);
@@ -14,10 +14,10 @@ const InputTodo = ({ addTodoItem }) => {
     e.preventDefault();
     if (title.trim()) {
       addTodoItem(title);
-      setTitle("");
-      setMessage("");
+      setTitle('');
+      setMessage('');
     } else {
-      setMessage("Please add a Todo");
+      setMessage('Please add a Todo');
     }
   };
 
@@ -31,7 +31,9 @@ const InputTodo = ({ addTodoItem }) => {
           onChange={handleChange}
           className="input-text"
         />
-        <button className="input-submit"><FaPlusCircle style={{ color: "#5e5e5e", fontSize: "22px" }}/></button>
+        <button type="button" className="input-submit">
+          <FaPlusCircle style={{ color: '#5e5e5e', fontSize: '22px' }} />
+        </button>
       </form>
       <span className="submit-warning">{message}</span>
     </>
